Moved Icons.vue into the directory where its siblings reside
