+++ src/components/home_sidebar_items/ComponentTab/Icons.vue
